+++ src/TwigExtension.php
@@ -11,7 +11,8 @@ namespace Zend\Expressive\Twig;
 
 use Twig_Extension;
 use Twig_SimpleFunction;
-use Zend\Expressive\Router\RouterInterface;
+use Zend\Expressive\Helper\ServerUrlHelper;
+use Zend\Expressive\Helper\UrlHelper;
 
 /**
  * Twig extension for rendering URLs and assets URLs from Expressive.
@@ -21,9 +22,14 @@ use Zend\Expressive\Router\RouterInterface;
 class TwigExtension extends Twig_Extension
 {
     /**
-     * @var RouterInterface
+     * @var ServerUrlHelper
      */
-    private $router;
+    private $serverUrlHelper;
+
+    /**
+     * @var UrlHelper
+     */
+    private $urlHelper;
 
     /**
      * @var string
@@ -36,18 +42,29 @@ class TwigExtension extends Twig_Extension
     private $assetsVersion;
 
     /**
-     * @param RouterInterface $router
-     * @param string $assetsUrl
-     * @param string $assetsVersion
+     * @var array
+     */
+    private $globals;
+
+    /**
+     * @param ServerUrlHelper $serverUrlHelper
+     * @param UrlHelper       $urlHelper
+     * @param string          $assetsUrl
+     * @param string          $assetsVersion
+     * @param array           $globals
      */
     public function __construct(
-        RouterInterface $router,
+        ServerUrlHelper $serverUrlHelper,
+        UrlHelper $urlHelper,
         $assetsUrl,
-        $assetsVersion
+        $assetsVersion,
+        array $globals = []
     ) {
-        $this->router        = $router;
-        $this->assetsUrl     = $assetsUrl;
-        $this->assetsVersion = $assetsVersion;
+        $this->serverUrlHelper = $serverUrlHelper;
+        $this->urlHelper       = $urlHelper;
+        $this->assetsUrl       = $assetsUrl;
+        $this->assetsVersion   = $assetsVersion;
+        $this->globals         = $globals;
     }
 
     /**
@@ -58,32 +75,76 @@ class TwigExtension extends Twig_Extension
         return 'zend-expressive';
     }
 
+    public function getGlobals()
+    {
+        return $this->globals;
+    }
+
     /**
      * @return Twig_SimpleFunction[]
      */
     public function getFunctions()
     {
         return [
-            new Twig_SimpleFunction('path', [$this, 'renderUri']),
+            new Twig_SimpleFunction('absolute_url', [$this, 'renderUrlFromPath']),
             new Twig_SimpleFunction('asset', [$this, 'renderAssetUrl']),
+            new Twig_SimpleFunction('path', [$this, 'renderUri']),
+            new Twig_SimpleFunction('url', [$this, 'renderUrl']),
         ];
     }
 
     /**
-     * Usage: {{ path('name', parameters) }}
+     * Render relative uri for a given named route
      *
-     * @param $name
-     * @param array $parameters
-     * @param bool $relative
+     * Usage: {{ path('article_show', {'id': '3'}) }}
+     * Generates: /article/3
+     *
+     * @param null  $route
+     * @param array $params
+     *
      * @return string
      */
-    public function renderUri($name, $parameters = [], $relative = false)
+    public function renderUri($route = null, $params = [])
     {
-        return $this->router->generateUri($name, $parameters);
+        return $this->urlHelper->generate($route, $params);
     }
 
     /**
+     * Render absolute url for a given named route
+     *
+     * Usage: {{ url('article_show', {'slug': 'article.slug'}) }}
+     * Generates: http://example.com/article/article.slug
+     *
+     * @param null  $route
+     * @param array $params
+     *
+     * @return string
+     */
+    public function renderUrl($route = null, $params = [])
+    {
+        return $this->serverUrlHelper->generate($this->urlHelper->generate($route, $params));
+    }
+
+    /**
+     * Render absolute url from a path
+     *
+     * Usage: {{ absolute_url('path/to/something') }}
+     * Generates: http://example.com/path/to/something
+     *
+     * @param $path
+     *
+     * @return string
+     */
+    public function renderUrlFromPath($path = null)
+    {
+        return $this->serverUrlHelper->generate($path);
+    }
+
+    /**
+     * Render asset url, optionally versioned
+     *
      * Usage: {{ asset('path/to/asset/name.ext', version=3) }}
+     * Generates: path/to/asset/name.ext?v=3
      *
      * @param $path
      * @param null $version